Moved plot_model into utilties subdirectory
