+++ model/prg_torch_model.py
@@ -20,8 +20,8 @@ from tensorflow.keras.preprocessing.image import load_img
 # load custom scripts
 from model.torch.VGG16_pretrained import VGG16_pretrained
 from model.torch.CustomDataset import CustomDataset
-from model.plot_model import plot_model_fit
 from model.torch.EarlyStopper import EarlyStopper
+from model.utilities.plot_model import plot_model_fit
 from model.utilities.plot_preds import plot_preds
 from model.utilities.plot_image import plot_image
 from model.utilities.plot_generator import plot_generator
